Final Push (For This Branch Most Likely)
It's been good working with you cv-core. I think we're done (sadly but fortunately). The live webcam shoplifting labeling system thing works. This is grand. I honestly didn't know how we'd do and we actually completed the core part of this project. I'm proud of the team. We have a bit to go, but the end is in site. :santa: